api/requirements.txt
dj_rest_auth==2.2.7
Django==4.1.5
djangorestframework==3.14.0
